MNIST/requirements.txt
Flask==1.0.2
h5py 
gevent
gunicorn
keras
numpy
Pillow==5.4.1
opencv-python
matplotlib==3.0.3
imageio
tensorflow==1.14.0
future
cassandra-driver
